move pointerinteraction from focus to pointer module
