+++ CHANGELOG.md
@@ -28,4 +28,5 @@
   a sorted list of entities and `HitData`, sorted by depth. This better reflects how pointer data
   would be expected to work, and makes it easier to find the topmost entity under each pointer.
 - Added: `get_nearest_hit` method added to `PointerInteraction`
+- Changed: Moved `PointerInteraction` from the `focus` module to `pointer`.
 - Added: viewport example
